sample-app/agent/src/pyproject.toml
[project]
name = "prism-task-assistant"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.4.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "bedrock-agentcore>=1.0.0",
]
